implemented migration

+++ main/models.py
@@ -9,58 +9,65 @@ class Board(models.Model):
     owner = models.ForeignKey(User, on_delete=models.CASCADE)
     category = models.CharField(max_length=200,default="New Board",blank=True)
     due_date = models.DateTimeField(null=True)
+    start_date = models.DateTimeField(null=True)
 
     def __str__(self):
         return self.category
     
-    def initialize_week(self,next_week=False):
-        today = timezone.localtime()
-        tz = today.tzinfo
-        year, week_num, day_of_week = today.isocalendar()
+    def initialize_week(self,given_datetime=False,next_week=False):
+        if not given_datetime:        
+            given_datetime = timezone.localtime()
+        tz = timezone.get_current_timezone()
+        year, week_num, day_of_week = given_datetime.isocalendar()
         if next_week:
             week_num+=1
-
+        self.start_date = tz.localize(datetime.fromisocalendar(year,week_num,1))
         for day_of_week in range(1,8):
-            day_datetime = tz.localize(datetime.fromisocalendar(year,week_num,day_of_week))
-            day_datetime = tz.localize(datetime.combine(day_datetime, datetime.max.time())) #sets time to 23:59 local time
-            name = f"{day_datetime.strftime('%A')} {day_datetime.date()}"
-            self.todolist_set.create(name=name,date=day_datetime)
+            dt = tz.localize(datetime.fromisocalendar(year,week_num,day_of_week))
+            date = dt.date()
+            name = f"{date.strftime('%A')} {date}"
+            self.todolist_set.create(name=name,date=date)
         
-        self.due_date = day_datetime
+        dt = tz.localize(datetime.combine(dt, datetime.max.time())) #sets dt to 23:59 local time
+        self.due_date = dt
         self.save()
 
-    def migrate_week(self,next_week=False):
+    def migrate_week(self,next_week=False,dt=None):
         week_todolists = self.todolist_set.exclude(date=None)
-        backlog = self.todolist_set.get(name="backlog")
-        futurelog = self.todolist_set.get(name="futurelog")
-        archive = self.todolist_set.get(name="archive")
 
-        #move complete tasks to archive, move incomplete tasks to backlog
-        archive(week_todolists,archive,backlog)
+        #all complete tasks in board are moved to archive, 
+        #tasks with due date up to datetime are moved to backlog
+        #if datetime=None then all incomplete tasks in board are moved to backlog
+        self.archive(week_todolists,datetime=dt)
+
+        #if not migrating to next week and if board is already current week
+        if not (next_week or self.start_date > timezone.now()):
+            return      
+       
+        given_datetime=False
+        if next_week:
+            given_datetime = datetime.combine(self.start_date, datetime.min.time())
         week_todolists.delete()
-        self.migrate_week(next_week)
-        #assign futurelog tasks due this week to weekdays
-        for task in futurelog.task_set.all():
-            for todolist in week_todolists:
-                if todolist.date == task.due_date:
-                    task.todolist = todolist
-                    task.save()
+        self.initialize_week(given_datetime=given_datetime,next_week=next_week)      
+        return
     
-    def archive(self,todolists,archive,backlog=None):
+    def archive(self,todolists,datetime=None):
+        if not datetime: datetime = self.due_date
+        archive = self.todolist_set.get(name="archive")
+        backlog = self.todolist_set.get(name="backlog")
         for todolist in todolists:
             for task in todolist.task_set.all():
                 if task.complete:
                     task.todolist = archive
-                else:
-                    if backlog:
-                        task.todolist = backlog
+                elif task.due_date and task.due_date.astimezone(tz=timezone.get_current_timezone()).date() < datetime.astimezone(tz=timezone.get_current_timezone()).date():
+                    task.todolist = backlog
                 task.save()
 
 
 class ToDoList(models.Model):
     board = models.ForeignKey(Board, on_delete=models.CASCADE)
     name = models.CharField(max_length=200,default="New List",blank=True)
-    date = models.DateTimeField(null=True)
+    date = models.DateField(null=True)
     position = models.IntegerField(default=0)
 
     class Meta:
